Move validateDiosphere to a better place
